refactor: move frontend components out of index page
